+++ spec/integration/shared_examples/unmanaged_device/delete.rb
@@ -1,0 +1,29 @@
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# You may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+
+RSpec.shared_examples 'UnmanagedDeviceDeleteExample' do |context_name|
+  include_context context_name
+
+  let(:item) { described_class.find_by(current_client, name: UNMANAGED_DEVICE1_NAME).first }
+
+  describe '#delete' do
+    it 'raises MethodUnavailable' do
+      expect { item.delete }.to raise_error(OneviewSDK::MethodUnavailable, /The method #delete is unavailable for this resource/)
+    end
+  end
+
+  describe '#remove' do
+    it 'remove resource' do
+      expect { item.remove }.to_not raise_error
+      expect(item.retrieve!).to eq(false)
+    end
+  end
+end