+++ CHANGELOG.md
@@ -9,12 +9,16 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Enclosure
    - Enclosure Group
    - Ethernet Network
+   - Fabric
    - FC Network
    - FCoE Network
    - Interconnect
+   - Internal Link Set
    - Logical Downlink
+   - Logical Enclosure
    - Logical Interconnect
    - Logical Interconnect Group
+   - Logical Switch
    - Logical Switch Group
    - Managed SANs
    - Network Set
@@ -25,7 +29,10 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - SAS Logical Interconnect Group
    - Scopes
    - Server Hardware Type
+   - Storage Pool
+   - Storage System
    - Switches
+   - Unmanaged Devices
    - User
 
 #### Breaking changes